UI: Reduce boilerplate for posting messages
Most of the methods in `puter.ui` do the same thing: generate a msg_id,
call postMessage() with that and some context information, and then set
a callback function for when it gets a response. The repetition
previously led to a bug where we missed out one of the fields.
(See 55370f760b7615b30de7faef5e950ed0d55bd122) So, let's extract the
boilerplate out.