UI: Support parent/child app connections
`puter.ui.launchApp()` now returns a `Promise<AppConnection>` to the
newly-launched child app, which is ready to receive messages when the
promise completes. This does mean that apps launched which do not
include puter.js will no longer resolve the promise.

`puter.ui.parentApp()` returns an `AppConnection` to the parent app, if
there is one.